// resources/js/enums/colors.ts
export enum ColorVariant {
    danger = 'error',
    danger_outline = 'error_outline',
    fuchsia = 'fuchsia',
    fuchsia_outline = 'fuchsia_outline',
    info = 'info',
    info_outline = 'info_outline',
    primary = 'primary',
    primary_outline = 'primary_outline',
    shade = 'shade',
    shade_outline = 'default_outline',
    success = 'success',
    success_outline = 'success_outline',
    warning = 'warning',
    warning_outline = 'warming_outline',
    white = 'white',
    transparent = 'transparent',
}